combined LLIELiv.vue and Upload.vue
