+++ frontend/src/components/LLIELib.vue
@@ -12,7 +12,32 @@
       </div>
 
       <div class="button">
-        <Upload @file-change="handleFileChange" @processed-image="handleProcessedImage"></Upload>
+        <div class="upload-container">
+          <label for="upload-input" class="upload-button">
+            选择文件
+            <input type="file" id="upload-input" @change="uploadImage" style="display: none;">
+          </label>
+        </div>
+
+        <div class="radios">
+          <el-radio-group v-model="cs">
+            <el-radio-button label="rgb" value="rgb" />
+            <el-radio-button label="hls" value="hls" />
+            <el-radio-button label="hsv" value="hsv" />
+            <el-radio-button label="lab" value="lab" />
+            <el-radio-button label="yuv" value="yuv" />
+          </el-radio-group>
+        </div>
+
+        <div class="radios">
+          <el-radio-group v-model="method">
+            <el-radio-button label="he" value="he" />
+            <el-radio-button label="clahe" value="clahe" />
+            <el-radio-button label="rclahe" value="rclahe" />
+          </el-radio-group>
+        </div>
+
+        <el-button type="primary" size="large" @click="processImage" :loading="processing">开始处理</el-button>
       </div>
     </div>
 
@@ -34,20 +59,71 @@
 </template>
 
 <script setup lang="ts">
+import api from '@/api/request.js';
+import { ElMessage } from 'element-plus';
 import { ref } from 'vue';
-import Upload from './Upload.vue';
 
 const originalImage = ref(null);
 const processedImage = ref(null);
+const method = ref(null);
+const file = ref(null);
+const cs = ref(null);
+const processing = ref(false);
 
-const handleFileChange = (data: string) => {
-  console.log('File changed:', new Blob([data]).size / 1024 / 1024, "MB");
-  originalImage.value = data;
+const emits = defineEmits(['processed-image', 'file-change'])
+
+const uploadImage = async (e) => {
+  try {
+    file.value = e.target.files[0];
+
+    if (file.value.type !== 'image/jpeg' && file.value.type !== 'image/png' && file.value.type !== 'image/jpg') {
+      ElMessage.error('请选择jpeg、png、jpg格式的图片文件')
+      return
+    }
+    else if (file.value.size > 1024 * 1024 * 8) {
+      ElMessage.error('文件大小超过限制(8MB)，请重新选择')
+      return
+    }
+
+    if (file) {
+      const reader = new FileReader();
+      reader.onload = function (e) {
+        originalImage.value = e.target.result;
+      };
+      reader.readAsDataURL(file.value);
+    } else {
+      ElMessage.error('上传文件不可为空')
+      return
+    }
+  } catch (error) {
+    console.error('Error uploading image:', error);
+  }
 };
 
-const handleProcessedImage = (data: string) => {
-  console.log('Processed image:', new Blob([data]).size / 1024 / 1024, "MB");
-  processedImage.value = data;
+const processImage = async () => {
+  if (originalImage.value == null || method.value == null || cs.value == null) {
+    console.log(originalImage.value, method.value, cs.value)
+    ElMessage.error('请选择图片文件和算法');
+    return;
+  }
+  try {
+    processing.value = true;
+    const formData = new FormData();
+    formData.append('file', file.value);
+    formData.append('method', method.value);
+    formData.append('cs', cs.value);
+
+    const response = await api.post('/img', formData, {
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      },
+    });
+    // console.log(response.data.data)
+    processedImage.value = response.data.data;
+    processing.value = false;
+  } catch (error) {
+    console.error('Error uploading image:', error);
+  }
 };
 
 // 创建一个临时的 <a> 标签下载处理后的图片
@@ -119,12 +195,34 @@ const refreshPage = () => {
 .image {
   max-width: 90%;
   height: auto;
+  border-radius: 8%;
 }
 
 /* 按钮和单选框样式 */
 .button,
 .radios {
   margin: 8px 0;
+}
+
+.upload-container {
+  position: relative;
+  display: inline-block;
+  margin-top: 8px;
+}
+
+.upload-button {
+  display: inline-block;
+  padding: 10px 19px;
+  background-color: #409EFF;
+  color: #fff;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+  font-size: 14px;
+}
+
+.upload-button:hover {
+  background-color: #66B1FF;
 }
 
 /* 媒体查询：手机 */
